requirements.txt
ccxt>=2.8.0
pandas>=1.5.0
python-telegram-bot>=20.0
python-dotenv>=0.21.0
ta>=0.10.2